Made a simple web app UI
I have used Flask web app module to develop this ui. It is simple, transparent and very flexible to use, so that we could add features easily in the future.

Main function:
I have brought the working directory address variable outside the main function as the specific functions in my program would not be able to fetch that without executing the main(CLI).
I have also added an extension on video category. Ie = 'webm' . The screen recordings are normally saved in this format, so it would be usefull.

About HTML:
I have used google webdesigner to create the webpage and used the jinja to render data inside the html.
Added a link to our github page in the web app .

I have also uploaded two screen shot of the running of the apps.

Thank You